# api/swagger.json
{"openapi":"3.0.0","info":{"title":"Reddit Clone API","version":"1.0.0","description":"API documentation for Reddit Clone"},"paths":{"/vote/{commentId}/{direction}":{"get":{"summary":"Endpoint for voting on a comment","parameters":[{"name":"commentId","in":"path","required":true,"description":"ID of the comment","schema":{"type":"string"}},{"name":"direction","in":"path","required":true,"description":"Direction of the vote (up or down)","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"example":true}}}}}},"/votes":{"post":{"summary":"Get vote information for multiple comments","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"commentsIds":{"type":"array","items":{"type":"string"}}}}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"example":{"commentsTotals":{"commentId1":5,"commentId2":-2},"userVotes":{"commentId1":1,"commentId2":-1}}}}}}}},"/register":{"post":{"summary":"Register a new user","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"201":{"description":"User registered successfully"},"500":{"description":"Internal server error"}}}},"/user":{"get":{"summary":"Get user information","responses":{"200":{"description":"Success","content":{"application/json":{"example":{"username":"exampleUser"}}}},"500":{"description":"Internal server error"}}}},"/login":{"post":{"summary":"Log in user","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"User logged in successfully"},"422":{"description":"Invalid username or password"}}}},"/logout":{"post":{"summary":"Log out user","responses":{"200":{"description":"User logged out successfully"}}}},"/comments":{"get":{"summary":"Get comments","parameters":[{"name":"search","in":"query","required":false,"description":"Search term for comments","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"example":["comment1","comment2"]}}},"500":{"description":"Internal server error"}}},"post":{"summary":"Create a new comment","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string"},"body":{"type":"string"},"parentId":{"type":"string"},"rootId":{"type":"string"}}}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"example":"newComment"}}},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/comments/root/{rootId}":{"get":{"summary":"Get root comments","parameters":[{"name":"rootId","in":"path","required":true,"description":"ID of the root comment","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"example":["comment1","comment2"]}}},"500":{"description":"Internal server error"}}}},"/comments/{id}":{"get":{"summary":"Get a comment by ID","parameters":[{"name":"id","in":"path","required":true,"description":"ID of the comment","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"example":"comment1"}}},"500":{"description":"Internal server error"}}}}},"components":{},"tags":[]}